+++ test/sync.js
@@ -1,4 +1,3 @@
-/* eslint-env node */
 /* eslint global-require: 0, func-names: 0, no-shadow: 0 */
 /* eslint-disable prefer-arrow-callback */